Refactor Key.ts
