+++ __tests__/Keys.test.ts
@@ -1,11 +1,5 @@
-import {
-    Ed25519PrivateKey,
-    Ed25519PublicKey,
-    generateMnemonic,
-    KeyMismatchException,
-    ThresholdKey
-} from "../src/Keys";
 import * as nacl from "tweetnacl";
+import {Ed25519PrivateKey, Ed25519PublicKey, generateMnemonic, KeyMismatchException, ThresholdKey} from "../exports";
 
 // key from hedera-sdk-java tests, not used anywhere
 const privKeyBytes = Uint8Array.of(-37, 72, 75, -126, -114, 100, -78, -40, -15, 44, -29, -64, -96, -23, 58, 11, -116, -50, 122, -15, -69, -113, 57, -55, 119, 50, 57, 68, -126, 83, -114, 16);
