+++ src/contract/ContractInfoQuery.ts
@@ -6,10 +6,10 @@ import { Query } from "../generated/Query_pb";
 import { grpc } from "@improbable-eng/grpc-web";
 import { Response } from "../generated/Response_pb";
 import { SmartContractService } from "../generated/SmartContractService_pb_service";
-import { Ed25519PublicKey } from "../Keys";
 import { ContractId, ContractIdLike, contractIdToProto, contractIdToSdk } from "../types/ContractId";
 import { AccountId, accountIdToSdk } from "../types/AccountId";
 import { timestampToDate } from "../types/Timestamp";
+import { Ed25519PublicKey } from "../crypto/Ed25519PublicKey";
 
 export type ContractInfo = {
     contractId: ContractId;
