+++ src/contract/ContractUpdateTransaction.ts
@@ -6,12 +6,12 @@ import { BaseClient } from "../BaseClient";
 import { SmartContractService } from "../generated/SmartContractService_pb_service";
 
 import { ContractUpdateTransactionBody } from "../generated/ContractUpdate_pb";
-import { PublicKey } from "../Keys";
 import { newDuration } from "../util";
 import { ContractIdLike, contractIdToProto } from "../types/ContractId";
 import { AccountId, accountIdToProto } from "../types/AccountId";
 import { FileIdLike, fileIdToProto } from "../types/FileId";
 import { dateToTimestamp, timestampToProto } from "../types/Timestamp";
+import { PublicKey } from "../crypto/PublicKey";
 
 export class ContractUpdateTransaction extends TransactionBuilder {
     private readonly _body: ContractUpdateTransactionBody;
