+++ src/file/FileCreateTransaction.ts
@@ -6,9 +6,9 @@ import { BaseClient } from "../BaseClient";
 
 import { FileService } from "../generated/FileService_pb_service";
 import { FileCreateTransactionBody } from "../generated/FileCreate_pb";
-import { Ed25519PublicKey } from "../Keys";
 import { KeyList } from "../generated/BasicTypes_pb";
 import { dateToTimestamp, timestampToProto } from "../types/Timestamp";
+import { Ed25519PublicKey } from "../crypto/Ed25519PublicKey";
 
 export class FileCreateTransaction extends TransactionBuilder {
     private readonly _body: FileCreateTransactionBody;
