+++ src/file/FileUpdateTransaction.ts
@@ -5,11 +5,11 @@ import { grpc } from "@improbable-eng/grpc-web";
 import { BaseClient } from "../BaseClient";
 
 import { FileService } from "../generated/FileService_pb_service";
-import { PublicKey } from "../Keys";
 import { KeyList } from "../generated/BasicTypes_pb";
 import { FileUpdateTransactionBody } from "../generated/FileUpdate_pb";
 import { FileIdLike, fileIdToProto } from "../types/FileId";
 import { dateToTimestamp, timestampToProto } from "../types/Timestamp";
+import { PublicKey } from "../crypto/PublicKey";
 
 export class FileUpdateTransaction extends TransactionBuilder {
     private readonly _body: FileUpdateTransactionBody;
